Moved some tests to a new test case file
